+++ src/samples/randoms_batch/randoms_batch_api/mod.rs
@@ -37,7 +37,7 @@ where
 {
     fn from(response: ThreadRequestResponse<RandomsBatch<P>>) -> Self {
         let ThreadRequestResponse::MessagePoolItem(result) = response else {
-                panic!("must be a response to a call to the element")
+                panic!("must be a response to a call to the pool item")
             };
         result
     }